Struct adaboost_bindings::adaboost::AdaBoost

pub struct AdaBoost {
    pub n_estimators: usize,
    pub weak_learners: Vec<WeakLearner>,
}

Representa o algoritmo AdaBoost que usa como classificadores fracos decision stumps.

Fields
n_estimators: usize

Número de classificadores fracos usados para fazer a predição.

weak_learners: Vec<WeakLearner>

Classificadores fracos usados para fazer a predição.

Implementations

impl AdaBoost

pub fn new(n_estimators: usize) -> AdaBoost

Cria um novo algoritmo AdaBoost.

Arguments

n_estimators - Número de classificadores fracos usados para fazer a predição.

Returns

Novo algoritmo AdaBoost.

pub fn fit(
&mut self,
py: Python<'_>,
x: PyReadonlyArray2<'_, i64>,
y: PyReadonlyArray1<'_, i64>
) -> PyResult<()>

Treina o algoritmo AdaBoost clássico.

Arguments

x - Array numpy com as features dos dados de treino.

y - Array numpy com os labels dos dados de treino.

pub fn predict(
&self,
py: Python<'_>,
x: PyReadonlyArray2<'_, i64>
) -> PyResult<Py<PyArray1<i64>>>

Faz a predição com base em um array numpy. Como no algoritmo clássico a predição é feita com base na votação de diferentes classificadores fracos, sendo o valor final 0 se a soma das diferentes predições ponderadas por alpha for negativa e 1 caso contrário.

Arguments

x - Array numpy com as features.

Returns

Array numpy com as predições.