Struct adaboost_bindings::weak_learner::WeakLearner

pub struct WeakLearner {
    pub feature_index: usize,
    pub threshold: i64,
    pub polarity: i64,
    pub error: f64,
    pub alpha: f64,
}

Representa um classificador fraco, no caso um decision stump. Escolhi os decision stumps pois foram eles que foram utilizados na aula de boosting, além disso eles são bastante simples.

Fields
feature_index: usize

Indice da feature que será utilizada como critério de decisão.

threshold: i64

Valor do threshold que para qual a feature será comparada. Se a feature for maior que o threshold, o dado será classificado como 1, caso contrário, será classificado como 0.

polarity: i64

Polaridade da decisão. Se for 1, o dado será classificado como 1 se for maior que o threshold, caso contrário, será classificado como 0. Se for -1, o dado será classificado como 1 se for menor que o threshold, caso contrário, será classificado como 0.

error: f64

Erro do classificador fraco.

alpha: f64

Peso do classificador fraco.

Implementations

impl WeakLearner

pub fn new() -> WeakLearner

Cria um novo classificador fraco vazio, incapaz de classificar qualquer dado.

Returns

Novo classificador fraco.

pub fn fit(&mut self, weighted_data: WeightedData)

Treina o classificador fraco com os dados passados. Nessa função os valores de feature_index, threshold, polarity são definidos. Esses são escolhidos como sendo a combinação de valores que minimiza o erro dos dados passados, essa busca é exaustiva e testa todas as possibilidades.

O erro é definido como a soma dos pesos das samples que foram classificadas incorretamente. A classificação é feita utilizando o feature_index, threshold e polarity definidos no treinamento. Ela é definida como 1 se a feature for maior que o threshold e a polaridade for 1 ou se a feature for menor que o threshold e a polaridade for -1. Caso contrário, a classificação é definida como 0.

Depois de ter os valores de feature_index, threshold, polarity definidos, o erro e o peso do classificador fraco são calculados.

Arguments

weighted_data - Samples que serão utilizadas para treinar o classificador fraco.

pub fn predict(&self, features: Vec<i64>) -> i64

Classifica um dado levando em consideração que o classificador fraco já foi treinado. A classificação é feita utilizando o feature_index, threshold e polarity definidos no treinamento. Ela é definida como 1 se a feature for maior que o threshold e a polaridade for 1 ou se a feature for menor que o threshold e a polaridade for -1. Caso contrário, a classificação é definida como 0.

Arguments

features - Vetor de features X do dado que será classificado.

Returns

Classificação do dado, 0 ou 1.

Panics

Se o classificador fraco não foi treinado.