Struct adaboost_bindings::weighted_data::WeightedData

pub struct WeightedData {
    pub samples: Vec<Sample>,
}

Guarda o conjunto de samples, que são os dados com seus pesos e labels.

Fields
samples: Vec<Sample>

Conjunto de samples.

Implementations

impl WeightedData

pub fn new(samples: Vec<Sample>) -> WeightedData

Cria um novo conjunto de samples.

Arguments

samples - Conjunto de samples.

Returns

Novo conjunto de samples.

pub fn computeWeightedErrorRate(&self, weak_learner: &WeakLearner) -> f64

Calcula o erro que um classificador fraco comete ao classificar as samples. O erro é definido como a soma dos pesos das samples que foram classificadas incorretamente.

Arguments

weak_learner - Classificador fraco.

Returns

Erro do classificador fraco.

pub fn updateWeights(&mut self, weak_learner: &WeakLearner)

Atualiza os pesos das samples de acordo com a predição do classificador fraco. Se a predição for correta, o peso da sample é multiplicado por e^(-alpha). Se a predição for incorreta, o peso da sample é multiplicado por e^(alpha). Os pesos são de todas as samples são normalizados para que a soma deles seja 1.