Struct adaboost_bindings::sample::Sample

pub struct Sample {
    pub features: Vec<i64>,
    pub label: i64,
    /* private fields */
}

Representa um dado, com suas features, label, e seu peso. As features são representadas por um vetor de inteiros, e o label só pode ser 1 ou -1.

Fields
features: Vec<i64>

Vetor de features X do dado.

label: i64

Label y do dado, que só pode ser 1 ou -1.

Implementations

impl Sample

pub fn new(features: Vec<i64>, label: i64, weight: f64) -> Sample

Cria um novo dado.

Arguments

features - Vetor de features X do dado.

label - Label y do dado.

weight - Peso do dado.

Returns

Novo dado.

pub fn getWeight(&self) -> f64

Retorna o peso do dado.

Returns

Peso do dado.

pub fn setWeight(&mut self, weight: f64)

Define o peso do dado.

Arguments

weight - Peso do dado.